app/requirements.txt
Django==5.2.11
asgiref>=3.8.1
sqlparse==0.5.4
debugpy==1.5.1
mysqlclient==2.1.0
gunicorn==23.0.0
django-userforeignkey~=0.5.0
django-import-export>=3.3.7
pillow>=10.4.0
pytesseract>=0.3.10
watchdog>=4.0.0

#certifi==2021.10.8
#cffi==1.15.0
#charset-normalizer==2.0.12
#cryptography==36.0.1
#defusedxml==0.7.1
django-allauth>=0.63.3
django-autocomplete-light>=3.9.2
django-crum>=0.7
django-filter>=25.1
django-formtools>=2.5.1
django-redis>=5.4.0
django-select2>=8.3.0
django-simple-history>=3.7.0
django-qr-code==4.2.0
idna==3.7
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==1.3.1
urllib3==2.6.3
openai>=1.40.0
python-dotenv>=1.0.1
plotly
numpy
pandas
matplotlib
seaborn
# Fuzzy matching libraries used for merge candidate discovery
python-Levenshtein>=0.25.0
rapidfuzz>=3.4.0

